Refactor deselect_last_selection
This commit refactor the command deselect_last_selection_and_refocus and
renames it to undo_find_under_expand. As the name suggests it aims to
undo the find_under_expand command and where not possible (due to lack
of knowledge about what was expanded last) tries to be sensible about
what selection to subtract. This is not considered done, and as indicated
in the code, the behavior might be changed slightly if it becomes
apparent that the behavior is less than optimal.