Fix bug with insert line above on first line
If we are on the first line of a file, the insert line above command
does not work. As there is no way to get the line number in this
context, we just revert the order of the insert newline and move
backwards commands, respectively, as this will achieve our desired
goal.